requirements.txt
fastapi
uvicorn
tensorflow
keras
pillow
numpy
python-multipart
